refactor: extract showReactionConfirmation to shared utility
Move duplicated reaction confirmation dialog logic from 5 files
into a shared utility at src/utils/showReactionConfirmation.ts
to comply with DRY principle (Rule of Three).
